fix: bug on navbar and bug on filter drawer
On page refresh, the selected page style disappeared. I modified the
component so that it automatically detects which path it is in versus
keeping the value in a state.

When the page resized from device to full with the drawer open, it was
not possible to scroll. Added an onResize event to deactive the drawer.